# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API DANAZ Pay",
        "description": "API pour la gestion des paiements CNPS, Banques et Entreprises.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Principal local"
        }
    ],
    "paths": {
...
        "/api/company/declarations": {
            "get": {
                "tags": [
                    "Espace Entreprise"
                ],
                "summary": "Lister l'historique des déclarations de l'entreprise",
                "description": "Retourne les déclarations de l'entreprise connectée avec possibilités de filtres.",
                "operationId": "companyListDeclarations",
                "parameters": [
                    {
                        "name": "bank_id",
                        "in": "query",
                        "description": "ID de la banque",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reference",
                        "in": "query",
                        "description": "Référence de la transaction",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mobile_reference",
                        "in": "query",
                        "description": "Référence Mobile Money",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (ex: 2026-03)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Date de début",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Date de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Espace Entreprise"
                ],
                "summary": "Initier une nouvelle déclaration",
                "description": "Crée une déclaration et envoie la preuve de paiement (PDF).",
                "operationId": "InitiateDeclaration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "period",
                                    "amount",
                                    "payment_mode"
                                ],
                                "properties": {
                                    "bank_id": {
                                        "description": "ID de la banque (requis sauf pour MoMo)",
                                        "type": "integer"
                                    },
                                    "reference": {
                                        "description": "Référence du paiement",
                                        "type": "string"
                                    },
                                    "mobile_reference": {
                                        "description": "Référence transaction MoMo",
                                        "type": "string"
                                    },
                                    "period": {
                                        "description": "Période concernée",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "amount": {
                                        "description": "Montant payé",
                                        "type": "number"
                                    },
                                    "payment_mode": {
                                        "type": "string",
                                        "enum": [
                                            "virement",
                                            "especes",
                                            "ordre_virement",
                                            "mobile_money",
                                            "orange_money"
                                        ]
                                    },
                                    "proof_pdf": {
                                        "description": "Preuve de paiement (PDF)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "Statut initial (submited par défaut)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement initié avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/declarations/{id}": {
            "put": {
                "tags": [
                    "Espace Entreprise"
                ],
                "summary": "Modifier une déclaration",
                "description": "Corriger une déclaration (généralement après un rejet). Utilisez POST avec _method=PUT si vous envoyez un fichier PDF.",
                "operationId": "EditDeclaration",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "payment_mode"
                                ],
                                "properties": {
                                    "bank_id": {
                                        "type": "integer"
                                    },
                                    "reference": {
                                        "type": "string"
                                    },
                                    "mobile_reference": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "payment_mode": {
                                        "type": "string",
                                        "enum": [
                                            "virement",
                                            "especes",
                                            "ordre_virement",
                                            "mobile_money",
                                            "orange_money"
                                        ]
                                    },
                                    "proof_pdf": {
                                        "description": "Nouveau fichier PDF (optionnel)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Déclaration corrigée avec succès"
                    },
                    "403": {
                        "description": "Impossible de modifier une transaction validée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/declarations/{id}/status": {
            "patch": {
                "tags": [
                    "Espace Entreprise"
                ],
                "summary": "Changer directement le statut (Utilitaire)",
                "description": "4. (Utilitaire) Changer directement le statut",
                "operationId": "changeDeclarationStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "initiated",
                                            "submited",
                                            "bank_validated",
                                            "cnps_validated",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/declarations/{id}/download-receipt": {
            "get": {
                "tags": [
                    "Espace Entreprise"
                ],
                "summary": "Télécharger la quittance officielle de la CNPS",
                "description": "Télécharge le fichier PDF de la quittance depuis le serveur FTP de la CNPS.",
                "operationId": "downloadReceipt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier de la quittance",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Quittance introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}